products api good

+++ db/migrate/20150816040147_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration
   def change
     create_table :products do |t|
       t.string :name, null: false
-      t.integer :efficiency, null: false
+      t.float :efficiency, null: false
 
       t.timestamps null: false
     end
